+++ CHANGELOG.md
@@ -38,6 +38,9 @@ This follows the new spec for the core SDKs that will come to swift shortly.
 - (Breaking): replace ropsten environment for sandbox
 Ropsten has been deprecated and won't work anymore. Sandbox is the preferred testing environment.
 
+- (Breaking): rename PrivateKey, PublicKey, KeyPair and CurvePoint to ECPrivateKey, ECPublicKey, ECKeyPair and ECCurvePoint respectively
+The previous names were too generic and would easily conflict with other classes/structs. These have then been prefixed with EC for Elliptic Curve.
+
 ### Fixed
 
 - re-include macos as a Cocoapods target